ユーザー登録、ログインAPIの脱device化。 - ユーザー登録時にbcrypt gemを使用 - 上記に伴いDB変更 - 既存のdeviceコード削除 - ルーティング見直し - Userモデルからdevice除去
