+++ app/models/user.rb
@@ -1,11 +1,11 @@
 class User < ApplicationRecord
-  # has_secure_password
+  has_secure_password
 
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
-  devise :database_authenticatable, :registerable,
-         :recoverable, :rememberable, :validatable,
-         :confirmable, :lockable
+  # devise :database_authenticatable, :registerable,
+  #        :recoverable, :rememberable, :validatable,
+  #        :lockable
 
   validates :nickname, presence: true
 
@@ -13,5 +13,8 @@ class User < ApplicationRecord
   has_one :yield_config
   has_one :retirement_asset_calc
 
-  # def password_digest; encrypted_password; end
+  def encrypted_password=(encrypted)
+    password_digest=(encrypted)
+  end
+  # def password_digest(password = nil); encrypted_password; end
 end
