+++ db/ridgepole.rb
@@ -25,7 +25,7 @@ create_table "users", force: :cascade do |t|
 
   ## Database authenticatable
   t.string "email", null: false, default: ""
-  t.string "encrypted_password", null: false, default: ""
+  t.string "password_digest", null: false, default: ""
 
   ## Recoverable
   t.string   "reset_password_token"
